Add isRouteDefined helper; split resolve.ts from client boundary
Extracts resolveRoute, isRouteDefined, CMSLink, and LinkHref into
link/resolve.ts (no "use client") so server components can call them
without crossing the client module boundary. link/index.tsx re-exports
all four for backwards compatibility.

isRouteDefined(link) should be used in place of !!link when guarding
on whether a CMS link has a real destination — the link field always
stores a partial initialValue object even when no destination is
selected, so object truthiness is never a reliable check.
